More Gui, Gui backend
further work implementing dialog. Linked up the first real functions.
Connects most buttons to variables and actions in the main dialog.
Implemented CRS selection in main screen, as well a passing this to the
plan drill trace writer. This needs to be rolled out to the other layer
writing code blocks (section crs will probably be hardcoded in). Minor
adjustments to GUI layout