Drilhole Filter
Implemented filtering drillholes using an envelope defined by the
section line and a +- width. This then intersects with the plan
drillhole traces, and any hole which goes through is included. Also
implemented this with the gui components, so the available/selected DH
lists update automatically. This can handle re running the filter,
adding more holes then running filter (sets it back to only holes in the
envelope) and change of the envelope resulting in a change in the
intersected drillholes closes #8