Section Generation
Implemented choosing section parameters by drawing on the map (closes
#10), also choosing section parameters from selected existing line
(closes #9). Implemented a preview of the section line and envelope
using rubberbands, and refreshes on changing of any of the section
parameters (closes #11) Implemented including the section facing in the
section definition file and is displayed in the Section View GUI. Most
error handling around this also done. Cause for the not refresh on load
and no legend display bugs found (blocking of signals while loading
sections) a small performance hit, although not sure if this is due to
the signal firing the visibility loop, or associated with generating the
legend settings fixes #7. Implemented some error catching to prevent
user from trying to create sections with incomplete parameters or
existing names fixes #19. This is communicated to the user through a
popup message bar in the Gen Sec GUI. Also cleaned out minor code
snippets printing to console for no longer needed debugging purposes