Measure Tools
Implemented a line measure and area measure tool closes #17, which
outputs to a shared dock widget in SectionView. Added to actions,
tolbars and exclusive with other map nav and feature tools. seems to be
working as intended, need to clean up some console print outs. was also
some playing around with resources to try and fix the loading error, no
success