Implemented negative dip handling
Implemented handling both positive and negative survey dips, using a
user selection at project setup. The negative/positive choice is
recorded in the save data and read from save also. Not fully tested, but
appears to work in that test which has been done. Still need to include
some routine in the data validation to flag when a sign contrary to the
user setup is chosen